Refactor GetPostalCodeData
