+++ app/Filament/Resources/ClientResource/Pages/CreateClient.php
@@ -3,32 +3,16 @@
 namespace App\Filament\Resources\ClientResource\Pages;
 
 use App\Filament\Resources\ClientResource;
-use App\Models\ClientAddress;
+use App\Filament\Resources\ClientResource\GetPostalCodeData;
 use Filament\Resources\Pages\CreateRecord;
-use Illuminate\Support\Str;
 
 class CreateClient extends CreateRecord
 {
+    use GetPostalCodeData;
+
     protected static string $resource = ClientResource::class;
 
     protected static ?string $title = 'Criar';
-
-    public function buttonClick(string $state, string $path)
-    {
-        $data = ClientAddress::GetAddress($state);
-
-        $id = explode('.', $path);
-
-        if ($data !== 404) {
-            $this->data['addresses'][$id[2]]['address_1'] = $data['street'];
-            $this->data['addresses'][$id[2]]['address_2'] = $data['neighborhood'];
-            $this->data['addresses'][$id[2]]['state_id'] = $data['state'];
-            $this->data['addresses'][$id[2]]['city_id'] = $data['city'];
-        } else {
-            $this->notify('danger', __('CEP não encontrado na base de dados do correios!'));
-        }
-        $this->notify('success', __('CEP válido!'));
-    }
 
     protected function getRedirectUrl(): ?string
     {
